requirements.api.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
opencv-python==4.9.0.80
numpy==1.26.4
onnxruntime>=1.24.1
paho-mqtt==2.1.0
pyyaml==6.0.1
python-multipart==0.0.9
